Change interfaces to be immutable
Replaces libfibre_subscribe_to_interface() by libfibre_get_interface_info()
and siblings.

The mutability of interfaces was intended to be used for cases like dynamic
lists or dictionaries where members are added and removed dynamically.
However it is more sensible and ergonomic if such mutability is provided on a
per-object basis rather than a per-interface basis.

We can add this once the low level protocol starts supporting it.